python/requirements.txt
flask
python-dotenv
stripe
mysql-connector-python
PyJWT